feat!: move equivalency to aweXpect.Core_net8.0.txt
Move the equivalency options and logic to aweXpect.Core, so that extensions can build on top of them.

Remove the `IsJsonSerializable` expectations on objects as they should be part of aweXpect.Json
