+++ Source/aweXpect.Core/Equivalency/EquivalencyComparisonType.cs
