+++ Source/aweXpect.Core/Equivalency/EquivalencyDefaults.cs
