+++ Source/aweXpect.Core/Equivalency/EquivalencyOptions.cs
