+++ Source/aweXpect.Core/Equivalency/EquivalencyTypeOptions.cs
