+++ Tests/aweXpect.Api.Tests/Expected/aweXpect_net8.0.txt
@@ -16,11 +16,6 @@ namespace aweXpect
         public static aweXpect.EnumerableQuantifier Exactly(int expected, aweXpect.Core.ExpectationGrammars expectationGrammars = 0) { }
         public static aweXpect.EnumerableQuantifier None(aweXpect.Core.ExpectationGrammars expectationGrammars = 0) { }
     }
-    public static class EquivalencyExtensions
-    {
-        public static TSelf Equivalent<TType, TThat, TElement, TSelf>(this aweXpect.Results.ObjectEqualityResult<TType, TThat, TElement, TSelf> result, System.Func<aweXpect.Equivalency.EquivalencyOptions, aweXpect.Equivalency.EquivalencyOptions>? options = null)
-            where TSelf : aweXpect.Results.ObjectEqualityResult<TType, TThat, TElement, TSelf> { }
-    }
     public static class ThatAsyncEnumerable
     {
         public static aweXpect.ThatAsyncEnumerable.Elements All(this aweXpect.Core.IThat<System.Collections.Generic.IAsyncEnumerable<string?>?> subject) { }
@@ -750,10 +745,6 @@ namespace aweXpect
         public static aweXpect.Results.AndOrResult<TSubject, aweXpect.Core.IThat<TSubject>> IsEquivalentTo<TSubject, TExpected>(this aweXpect.Core.IThat<TSubject> source, TExpected expected, System.Func<aweXpect.Equivalency.EquivalencyOptions<TExpected>, aweXpect.Equivalency.EquivalencyOptions>? options = null, [System.Runtime.CompilerServices.CallerArgumentExpression("expected")] string doNotPopulateThisValue = "") { }
         public static aweXpect.Results.AndOrResult<object?, aweXpect.Core.IThat<object?>> IsExactly(this aweXpect.Core.IThat<object?> source, System.Type type) { }
         public static aweXpect.Results.AndOrWhoseResult<TType, aweXpect.Core.IThat<object?>> IsExactly<TType>(this aweXpect.Core.IThat<object?> source) { }
-        public static aweXpect.Results.AndOrResult<object?, aweXpect.Core.IThat<object?>> IsJsonSerializable(this aweXpect.Core.IThat<object?> source, System.Func<aweXpect.Equivalency.EquivalencyOptions, aweXpect.Equivalency.EquivalencyOptions>? equivalencyOptions = null) { }
-        public static aweXpect.Results.AndOrResult<object?, aweXpect.Core.IThat<object?>> IsJsonSerializable(this aweXpect.Core.IThat<object?> source, System.Text.Json.JsonSerializerOptions serializerOptions, System.Func<aweXpect.Equivalency.EquivalencyOptions, aweXpect.Equivalency.EquivalencyOptions>? equivalencyOptions = null) { }
-        public static aweXpect.Results.AndOrResult<object?, aweXpect.Core.IThat<object?>> IsJsonSerializable<T>(this aweXpect.Core.IThat<object?> source, System.Func<aweXpect.Equivalency.EquivalencyOptions, aweXpect.Equivalency.EquivalencyOptions>? equivalencyOptions = null) { }
-        public static aweXpect.Results.AndOrResult<object?, aweXpect.Core.IThat<object?>> IsJsonSerializable<T>(this aweXpect.Core.IThat<object?> source, System.Text.Json.JsonSerializerOptions serializerOptions, System.Func<aweXpect.Equivalency.EquivalencyOptions, aweXpect.Equivalency.EquivalencyOptions>? equivalencyOptions = null) { }
         public static aweXpect.Results.AndOrResult<object?, aweXpect.Core.IThat<object?>> IsNot<TType>(this aweXpect.Core.IThat<object?> source) { }
         public static aweXpect.Results.AndOrResult<T?, aweXpect.Core.IThat<T?>> IsNot<T>(this aweXpect.Core.IThat<T?> source, System.Type type)
             where T :  class { }
@@ -865,36 +856,10 @@ namespace aweXpect
 }
 namespace aweXpect.Equivalency
 {
-    public static class EquivalencyAwexpectCustomizationExtensions
+    public static class EquivalencyExtensions
     {
-        public static aweXpect.Equivalency.EquivalencyAwexpectCustomizationExtensions.EquivalencyCustomization Equivalency(this aweXpect.Customization.AwexpectCustomization awexpectCustomization) { }
-        public class EquivalencyCustomization : aweXpect.Customization.ICustomizationValueUpdater<aweXpect.Equivalency.EquivalencyAwexpectCustomizationExtensions.EquivalencyCustomizationValue>
-        {
-            public aweXpect.Customization.ICustomizationValueSetter<aweXpect.Equivalency.EquivalencyOptions> DefaultEquivalencyOptions { get; }
-            public aweXpect.Equivalency.EquivalencyAwexpectCustomizationExtensions.EquivalencyCustomizationValue Get() { }
-            public aweXpect.Customization.CustomizationLifetime Update(System.Func<aweXpect.Equivalency.EquivalencyAwexpectCustomizationExtensions.EquivalencyCustomizationValue, aweXpect.Equivalency.EquivalencyAwexpectCustomizationExtensions.EquivalencyCustomizationValue> update) { }
-        }
-        public class EquivalencyCustomizationValue : System.IEquatable<aweXpect.Equivalency.EquivalencyAwexpectCustomizationExtensions.EquivalencyCustomizationValue>
-        {
-            public EquivalencyCustomizationValue() { }
-            public aweXpect.Equivalency.EquivalencyOptions DefaultEquivalencyOptions { get; init; }
-        }
-    }
-    public enum EquivalencyComparisonType
-    {
-        ByValue = 0,
-        ByMembers = 1,
-    }
-    public static class EquivalencyDefaults
-    {
-        public static aweXpect.Equivalency.EquivalencyComparisonType DefaultComparisonType(System.Type type) { }
-    }
-    public class EquivalencyOptions : aweXpect.Equivalency.EquivalencyTypeOptions, System.IEquatable<aweXpect.Equivalency.EquivalencyOptions>
-    {
-        public EquivalencyOptions() { }
-        public System.Collections.Generic.Dictionary<System.Type, aweXpect.Equivalency.EquivalencyTypeOptions> CustomOptions { get; init; }
-        public System.Func<System.Type, aweXpect.Equivalency.EquivalencyComparisonType> DefaultComparisonTypeSelector { get; init; }
-        public aweXpect.Equivalency.EquivalencyOptions For<TMember>(System.Func<aweXpect.Equivalency.EquivalencyTypeOptions, aweXpect.Equivalency.EquivalencyTypeOptions> options) { }
+        public static TSelf Equivalent<TType, TThat, TElement, TSelf>(this aweXpect.Results.ObjectEqualityResult<TType, TThat, TElement, TSelf> result, System.Func<aweXpect.Equivalency.EquivalencyOptions, aweXpect.Equivalency.EquivalencyOptions>? options = null)
+            where TSelf : aweXpect.Results.ObjectEqualityResult<TType, TThat, TElement, TSelf> { }
     }
     public static class EquivalencyOptionsExtensions
     {
@@ -906,28 +871,6 @@ namespace aweXpect.Equivalency
             where TEquivalencyOptions : aweXpect.Equivalency.EquivalencyTypeOptions { }
         public static TEquivalencyOptions IncludingProperties<TEquivalencyOptions>(this TEquivalencyOptions @this, aweXpect.Equivalency.IncludeMembers propertiesToInclude = 2)
             where TEquivalencyOptions : aweXpect.Equivalency.EquivalencyTypeOptions { }
-    }
-    public class EquivalencyOptions<TExpected> : aweXpect.Equivalency.EquivalencyOptions, System.IEquatable<aweXpect.Equivalency.EquivalencyOptions<TExpected>>
-    {
-        public EquivalencyOptions(aweXpect.Equivalency.EquivalencyOptions inner) { }
-        public new aweXpect.Equivalency.EquivalencyOptions<TExpected> For<TMember>(System.Func<aweXpect.Equivalency.EquivalencyTypeOptions, aweXpect.Equivalency.EquivalencyTypeOptions> options) { }
-    }
-    public class EquivalencyTypeOptions : System.IEquatable<aweXpect.Equivalency.EquivalencyTypeOptions>
-    {
-        public EquivalencyTypeOptions() { }
-        public aweXpect.Equivalency.EquivalencyComparisonType? ComparisonType { get; init; }
-        public aweXpect.Equivalency.IncludeMembers Fields { get; init; }
-        public bool IgnoreCollectionOrder { get; init; }
-        public string[] MembersToIgnore { get; init; }
-        public aweXpect.Equivalency.IncludeMembers Properties { get; init; }
-    }
-    [System.Flags]
-    public enum IncludeMembers
-    {
-        None = 0,
-        Public = 2,
-        Internal = 4,
-        Private = 8,
     }
 }
 namespace aweXpect.Options
