+++ Tests/aweXpect.Internal.Tests/Helpers/EquivalencyComparerTests.ComparisonTypeTests.cs
@@ -2,7 +2,7 @@
 
 // ReSharper disable UnusedAutoPropertyAccessor.Local
 
-namespace aweXpect.Internal.Tests.Equivalency;
+namespace aweXpect.Internal.Tests.Helpers;
 
 public sealed partial class EquivalencyComparerTests
 {
