+++ Tests/aweXpect.Internal.Tests/Helpers/EquivalencyComparerTests.CustomTypeTests.cs
@@ -2,7 +2,7 @@
 
 // ReSharper disable NotAccessedPositionalProperty.Local
 
-namespace aweXpect.Internal.Tests.Equivalency;
+namespace aweXpect.Internal.Tests.Helpers;
 
 public sealed partial class EquivalencyComparerTests
 {
