+++ Tests/aweXpect.Internal.Tests/Helpers/EquivalencyComparerTests.PropertyTests.cs
@@ -1,6 +1,6 @@
 ﻿using aweXpect.Equivalency;
 
-namespace aweXpect.Internal.Tests.Equivalency;
+namespace aweXpect.Internal.Tests.Helpers;
 
 public sealed partial class EquivalencyComparerTests
 {
