+++ Tests/aweXpect.Tests/Collections/ThatAsyncEnumerable.DoesNotStartWith.Tests.cs
@@ -1,5 +1,6 @@
 ﻿#if NET8_0_OR_GREATER
 using System.Collections.Generic;
+using aweXpect.Equivalency;
 
 // ReSharper disable PossibleMultipleEnumeration
