+++ Tests/aweXpect.Tests/Collections/ThatEnumerable.DoesNotStartWith.Tests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using aweXpect.Equivalency;
 
 // ReSharper disable PossibleMultipleEnumeration
